@@ -2,9 +2,10 @@ package com.lowbottgames.au.sydney.traffic.cam.fragment;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v7.preference.Preference;
-import android.support.v7.preference.PreferenceFragmentCompat;
-import android.support.v7.preference.PreferenceManager;
+
+import androidx.preference.Preference;
+import androidx.preference.PreferenceFragmentCompat;
+import androidx.preference.PreferenceManager;
 
 import com.lowbottgames.au.sydney.traffic.cam.BuildConfig;
 import com.lowbottgames.au.sydney.traffic.cam.R;
