@@ -1,16 +1,17 @@
 package com.lowbottgames.au.sydney.traffic.cam;
 
 import android.content.SharedPreferences;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.preference.PreferenceManager;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.preference.PreferenceManager;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
 import com.lowbottgames.au.sydney.traffic.cam.adapter.PlaceItemsRVAdapter;
 import com.lowbottgames.au.sydney.traffic.cam.domain.CamItem;
@@ -75,7 +76,7 @@ public class MainActivity extends AppCompatActivity {
         if (placeItems != null) {
             for (PlaceItem placeItem: placeItems) {
                 for (CamItem camItem: placeItem.camItem) {
-                    Picasso.with(this).invalidate(TCSHelper.getImageURLString(camItem.camID));
+                    Picasso.get().invalidate(TCSHelper.getImageURLString(camItem.camID));
                 }
             }
         }
