@@ -1,14 +1,15 @@
 package com.lowbottgames.au.sydney.traffic.cam.adapter;
 
 import android.content.Context;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.lowbottgames.au.sydney.traffic.cam.R;
 import com.lowbottgames.au.sydney.traffic.cam.domain.CamItem;
